@@ -1,12 +1,24 @@
 package com.techstudy.moviebookingapp.controller;
 
-import com.techstudy.moviebookingapp.model.*;
+
+import com.techstudy.moviebookingapp.exceptions.InvalidInputException;
+import com.techstudy.moviebookingapp.model.BookingDetails;
+import com.techstudy.moviebookingapp.model.MovieDescription;
+import com.techstudy.moviebookingapp.model.MovieDetails;
+import com.techstudy.moviebookingapp.model.PopularMovieResponse;
 import com.techstudy.moviebookingapp.serviceimpl.MovieServiceImpl;
+import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
-import javax.validation.Valid;
 import java.util.List;
 
 /**
@@ -28,10 +40,14 @@ public class MovieController {
      * Method to get the first 10 popular movies
      * @return List<MovieDescription
      */
-    @GetMapping("/popular/movies")
-    public ResponseEntity<List<MovieDescription>> getPopularMovies(){
-        List<MovieDescription> popularMovieResponse =  serviceImpl.getPopularMovies();
-        return new ResponseEntity<>(popularMovieResponse, HttpStatus.OK);
+    @GetMapping("/popular/movies/{number}")
+    public ResponseEntity<List<MovieDescription>> getPopularMovies(@PathVariable String number){
+        try{
+            List<MovieDescription> popularMovieResponse =  serviceImpl.getPopularMovies(number);
+            return new ResponseEntity<>(popularMovieResponse, HttpStatus.OK);
+        }catch (NumberFormatException e) {
+            throw new InvalidInputException("Kindly provide the valid number in the popularNumber");
+        }
     }
 
     /**
@@ -40,7 +56,7 @@ public class MovieController {
      * @return PopularMovieResponse
      */
     @GetMapping("/search/movies")
-    public ResponseEntity<PopularMovieResponse> searchMovies(@RequestParam(required = true) String input){
+    public ResponseEntity<PopularMovieResponse> searchMovies(@RequestParam String input){
         PopularMovieResponse searchResponse =  serviceImpl.searchMovie(input);
         return new ResponseEntity<>(searchResponse, HttpStatus.OK);
     }
@@ -51,7 +67,7 @@ public class MovieController {
      * @return MovieDetails
      */
     @GetMapping("/movies/{movie_id}")
-    public ResponseEntity<MovieDetails> getMovieDetails(@PathVariable(name = "movie_id", required = true) String movieId){
+    public ResponseEntity<MovieDetails> getMovieDetails(@PathVariable(name = "movie_id") String movieId){
         MovieDetails movieDetails =  serviceImpl.getMovieDetails(movieId);
         return new ResponseEntity<>(movieDetails, HttpStatus.OK);
     }
@@ -61,8 +77,8 @@ public class MovieController {
      * @param bookingDetails the booking details
      * @return BookingDetails
      */
-    @PostMapping("/movie/bookTicket")
-    public ResponseEntity<BookingDetails> createBooking(@RequestBody @Valid BookingDetails bookingDetails) {
+    @PostMapping("/bookTicket")
+    public ResponseEntity<BookingDetails> createBooking(@RequestBody @Valid  BookingDetails bookingDetails) {
         BookingDetails bookings = serviceImpl.createBooking(bookingDetails);
         return new ResponseEntity<>(bookings, HttpStatus.CREATED);
     }
