@@ -7,23 +7,28 @@ import com.techstudy.moviebookingapp.model.BookingDetails;
 import com.techstudy.moviebookingapp.model.BookingTicketDetails;
 import com.techstudy.moviebookingapp.repository.AdminRepository;
 import com.techstudy.moviebookingapp.service.AdminService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import java.util.Date;
 import java.util.List;
 
 /**
  * Service implementation class for Admin related functionalities
+ *
  * @author souvikdey
  */
 @Service
 public class AdminServiceImpl implements AdminService {
 
-    @Autowired
-    private AdminRepository adminRepository;
+    private final AdminRepository adminRepository;
+
+    public AdminServiceImpl(AdminRepository adminRepository) {
+        this.adminRepository = adminRepository;
+    }
 
     /**
      * Business implementation for viewing all the bookings
+     *
      * @return List<BookingDetails>
      */
     @Override
@@ -33,15 +38,16 @@ public class AdminServiceImpl implements AdminService {
 
     /**
      * Business implementation for cancel the booking by providing an email id
-     * @param email
+     *
+     * @param email email id needs to be provided
      * @return BookingDetails
      */
     @Override
-    public BookingDetails cancelBooking(String email){
+    public BookingDetails cancelBooking(String email) {
 
         BookingDetails updatedBooking = null;
         BookingDetails existingBookingDetails = adminRepository.findByEmail(email);
-        if(existingBookingDetails != null){
+        if (existingBookingDetails != null) {
             updatedBooking = new BookingDetails();
             updatedBooking.setBookingStatus("cancelled");
             updatedBooking.setNumberOfTickets(0);
@@ -56,7 +62,8 @@ public class AdminServiceImpl implements AdminService {
 
     /**
      * Business implementation for searching a booking by the admin input which will match with firstname, lastname & email
-     * @param input
+     *
+     * @param input input string needs to be provided
      * @return List<BookingDetails>
      */
     @Override
@@ -66,14 +73,15 @@ public class AdminServiceImpl implements AdminService {
 
     /**
      * Business implementation for searching a booking by passing fromdate & todate
-     * @param fromDate
-     * @param toDate
+     *
+     * @param fromDate from date
+     * @param toDate   to date
      * @return List<BookingDetails>
      */
     @Override
     public List<BookingDetails> filterBooking(Date fromDate, Date toDate) {
-        List<BookingDetails> bookingDetails = adminRepository.filterBooking(fromDate,toDate);
-        if(bookingDetails.size() == 0){
+        List<BookingDetails> bookingDetails = adminRepository.filterBooking(fromDate, toDate);
+        if (bookingDetails.isEmpty()) {
             throw new FilterBookingException();
         }
         return bookingDetails;
@@ -81,26 +89,26 @@ public class AdminServiceImpl implements AdminService {
 
     /**
      * Business implementation for updating an existing booking
-     * @param email
-     * @param bookingTicketDetails
+     *
+     * @param email                email id
+     * @param bookingTicketDetails booking details
      * @return BookingDetails
-     * @throws Exception
      */
     @Override
-    public BookingDetails updateBooking(String email, BookingTicketDetails bookingTicketDetails) throws Exception {
+    public BookingDetails updateBooking(String email, BookingTicketDetails bookingTicketDetails) {
         int maxTicketCount = 10;
         BookingDetails existingBookingDetails = adminRepository.findByEmail(email);
-        if(existingBookingDetails != null){
-            if(bookingTicketDetails.getBookingDate() != null){
+        if (existingBookingDetails != null) {
+            if (bookingTicketDetails.getBookingDate() != null) {
                 existingBookingDetails.setBookingDate(bookingTicketDetails.getBookingDate());
             }
-            if(bookingTicketDetails.getNumberOfTickets() != 0 && maxTicketCount >= bookingTicketDetails.getNumberOfTickets()){
+            if (bookingTicketDetails.getNumberOfTickets() != 0 && maxTicketCount >= bookingTicketDetails.getNumberOfTickets()) {
                 existingBookingDetails.setNumberOfTickets(bookingTicketDetails.getNumberOfTickets());
-            }else{
+            } else {
                 throw new UpdateBookingWithMaxLimitException(email);
             }
             return adminRepository.save(existingBookingDetails);
-        }else{
+        } else {
             throw new UpdateBookingException();
         }
     }
