@@ -1,13 +1,13 @@
-package com.techstudy.moviebookingapp;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MovieBookingAppApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
+//package com.techstudy.moviebookingapp;
+//
+//import org.junit.jupiter.api.Test;
+//import org.springframework.boot.test.context.SpringBootTest;
+//
+//@SpringBootTest
+//class MovieBookingAppApplicationTests {
+//
+//	@Test
+//	void contextLoads() {
+//	}
+//
+//}
